stm32f: update flash write size
This CL updates the constant FLASH_WRITE_BYTES to 64 and modifies the
decrement value in the flash_physical_write() loop.

The constant value was changed so that it would agree with
EC_FLASH_SIZE_MAX which is the size of the data payload.
FLASH_WRITE_BYTES is provided as part of EC_CMD_FLASH_INFO, so
programs which query size this way will can take advantage of 64-byte
writes.

The decrement value in the loop did not actually change in value, but
instead uses the size of the half-word.

Signed-off-by: David Hendricks <dhendrix@chromium.org>

BUG=none
TEST=Tested on Lucas

Change-Id: If335bd8e11db0acc6464dcdef819d91f61ae0890
(cherry picked from commit ef7557b4fd1a6a889d9918b2811567af216ebaf3)