Replacing usleep with interrupt friendly udelay
There was a usleep put into a function that can be called from in an
interrupt context, which doesn't work.  This just switches it over to
udelay which will work in an interrupt.

Also flips the condition on the i2c if/else that might send it there.
It was backwards before

BUG=chrome-os-partner:12688
TEST=Run "battery" "pmu" boot the machine and use the keyboard.  Then
replace the in_interrupt_context() function with "1" to force it to use
polling and repeat the test.  Everything should work in both cases.
BOARD=snow

Original-Change-Id: Ib2a8c7f9e5e2eb3f6b00678d6307afc9dd5f0518
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
(cherry picked from commit 4ac2b35ba515b3b517b32a04586d81a25fefef77)

Change-Id: Ie866871cafaa3d990350c4b31590c3d65f872e5c
Reviewed-on: https://gerrit.chromium.org/gerrit/30702
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Charlie Mooney <charliemooney@chromium.org>