link: Don't print keyboard state changes by default
The keyboard scan module generates a lot of debug output when the user
is typing on the keyboard - enough so that switching to the console,
logging in as root, and typing 'ectool console' flushes the EC's
console output buffer of any useful data.  Default printing this to
off.  Add a new 'ksstate' command which will print the current
keyboard scan state or toggle printing off/on.

This is important for debugging LPC communication failures.

BUG=chrome-os-partner:13819
BRANCH=link
TEST=manual

- Boot system
- Type on keyboard.  At EC console, should not see KB state: output
- Hold down space bar.
- At EC console, type 'ksstate'.  Should print:
  ksstate
  [20.943886 KB debounced : -- -- -- -- -- -- -- -- -- -- -- 20 --]
  [20.945215 KB prev      : -- -- -- -- -- -- -- -- -- -- -- 20 --]
  [20.945568 KB debouncing: -- -- -- -- -- -- -- -- -- -- -- -- --]
  Keyboard scan state printing off
- Release space bar
- At EC console, type 'ksstate on'
- Type on keyboard.  Should see KB state: output
- At EC console, type 'ksstate off'
- Type on keyboard.  At EC console, should not see KB state: output

Original-Change-Id: I4343b7b777fd13057b3222eeba77ed099c5e5a93
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35843
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
(cherry picked from commit 90a18599bb3ee0b0f49617a950805eb6bb77dd3f)

Change-Id: I512e85bb6afcd5b27a2ad74d7f60930f228a6da8
Reviewed-on: https://gerrit.chromium.org/gerrit/35898
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Tested-by: Randall Spangler <rspangler@chromium.org>