WORKAROUND: always detect special key combinations
When the fake battery key is asserted, the special runtime key
combination detection is counting the number of keys and getting 4
keys instead of 3.
In order to get the proper action in that case, relax the condition.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:18919
TEST=on Spring, press alt+VolUp+R with and without charger plugged and
see the machine rebooting in both cases.

Change-Id: I6dc0787f48aa3802f1f2804be6ee50a013b35575
Reviewed-on: https://gerrit.chromium.org/gerrit/49187
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>