spring: Shorten charging state update interval
The condition for the battery to set FULLY_CHARGED bit is:
  1. Charge >= 95%
  2. V >= 12.3V
  3. I < 52mA for 60 seconds

If we hit fastcharge timeout when 1 and 2 are true, the EC has to
respond and restart charging within 60 seconds. Otherwise, the battery
sets FULLY_CHARGED bit and stops charging with charge < 100%.

Let's shorten the update interval for S3/S5 from 60 seconds to 30
seconds. With current 10-second interval for charging state, the EC will
restart charging within 40 seconds.

BUG=chrome-os-partner:21093
TEST=Build success
BRANCH=Spring

Change-Id: I61c2f47af5101b91d5fea43db4e03c807f43ebc0
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62471
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>