Add an enum for the number of flash regions
Add an enum for the number of flash regions so we can keep track of all
the possible regions.

(This is used in U-Boot which wants to declare an array of all possible
regions.)

BUG=chromium:244019
BRANCH=none
TEST=manual
Build EC for pit.

Original-Change-Id: I494d857f1388dcc5c64b1cd580cf0084bdef0212
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62701
Reviewed-by: Randall Spangler <rspangler@chromium.org>
(cherry picked from commit db347953e1a463abcb50669988b3f45bd7fbb826)

Change-Id: I6d07f0f37dfe59122c2e95a08f8dcb24be7050e5
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63026