spring: Do not charge when cutting off the battery
If a charger is plugged in while we are cutting off the battery, the EC
will be powered by external power, and the battery will be in a state
where it doesn't charge. Let's stop charging in this case to get the
battery into cut-off state properly and resume charging 5 seconds after.

BUG=chrome-os-partner:21107
TEST=Plug in the charger while cutting off the battery.
BRANCH=Spring

Change-Id: I5505485be7872f9887efc725a2b3aa610c6b42b9
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63647
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>