spring: Workaround a bug that OTG dongle removal is not detected
TSU6721 doesn't report OTG dongle removal in S5. This prevents us from
properly detecting subsequent charger type. Since we don't supply VBUS
power in S5 for OTG dongle, if we see an OTG dongle with VBUS power in
S5, we can conclude this is a charger plugged in after an OTG dongle is
removed. By resetting TSU6721 to force a redetection in this case, we
can work around this bug.

This, however, consumes more power when a charger with ID=0 is plugged
in, as the EC would keep resetting TSU6721.

BUG=chrome-os-partner:21675
TEST=In S5, plug in OTG dongle and then unplug. Plug in a charger, and
see yellow LED.
TEST=In S5, plug in OTG dongle and then unplug. Boot the system and then
plug in a charger. See yellow LED.
BRANCH=Spring

Change-Id: I6546b629c748efc67bf28ee8751513bb7ac9d206
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64786
Reviewed-by: Todd Broch <tbroch@chromium.org>