Add abstract "cond_t" type to detect state transitions.
We often need to watch for transitions between one state and another, so
that we can issue warnings or take action ONCE. This abstracts that "have I
already reacted to this" stuff into a single set of functions.

For example, this code reads a GPIO every time through the loop, but it only
generates an event when the GPIO value changes from 0 to 1:

    cond_t c;

    cond_init_false(&c);

    while(1) {
        int val = read_some_gpio();
        cond_set(&c, val);

        if (cond_went_true(&c))
            host_event(SOMETHING_HAPPENED);

        sleep(1);
    }

BUG=none
BRANCH=falco,peppy
TEST=manual

Note: This doesn't really need cherry-picking from ToT (it was a
prerequisite for another CL that we decided NOT to backport), but it doesn't
hurt. Other than the tests, nothing uses this. It's easier to just pull it
in than to keep treating it as an exception.

To test:

  make BOARD=falco runtests

Original-Change-Id: I42393fcf3c4eb71b9551118a0f442d55c0691315
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65071
(cherry picked from commit 0c8c2e453ab3960f315050fbb9808f438398624f)

Change-Id: I44fe9a660016f3aa2ea2c7f8eb5c3e698d8f618f
Reviewed-on: https://gerrit.chromium.org/gerrit/66838
Commit-Queue: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Tested-by: Bill Richardson <wfrichar@chromium.org>