Handle multiple independent sources for CPU throttling
When both the thermal task and the power/charger task have different ideas
on whether the CPU should be throttled, we need to OR their opinions so that
either can start throttling, but both have to agree to stop it.

This CL changes the chipset_throttle_cpu() function to also take a "source"
argument, so we can handle multiple opinions.

As it turns out, this same also problem existed in the Falco-specific power
logic by itself, but was largely masked by the threshold settings. We handle
that now, too.

BUG=chrome-os-partner:20739
BRANCH=falco, ToT
TEST=manual

This change adds a bunch of tests to ensure that all this works, so try

  cd src/platform/ec
  make BOARD=falco runtests

And of course, try it on the actual hardware under the appropriate loads.

Change-Id: Id651e2cdc8dbd435aad6e21c5d2ce4b932a55f88
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168088
Reviewed-by: Randall Spangler <rspangler@chromium.org>