BQ24715: Disable 40kHz audio frequency limit option
The audio frequency limit function requires additional power
to keep the FET swtiching frequency above 40kHz. On some
systems it is also reported to cause additional audible noise.

Although the power-on setting is to disable this feature, we
need to ensure the charge controller options are updated from
running earlier versions of the driver.

BUG=chrome-os-partner:27063
BRANCH=None
TEST=Verify bit 10 is not set in the charger option register
using the 'charger' EC console command.

Change-Id: Ifab16ef3b53e25b5e82225040e7ab315d8760986
Signed-off-by: Dave Parker <dparker@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191770
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>