Kip: Modify battery LED behavior.
We won't to see the corner case when plugging in adpater.
if battery want to charge, we want to see the amber led directly
when plugging in adapter, not white to amber.
if battery into fully charged, we want to see the white led directly
when plugging in adapter, not amber to white.

BUG=None
BRANCH=rambi
TEST=manual
  Check battery LED show amber directly when plugging in adpater,
  if battery needs to charge.
  Check battery LED show white directly when plugging in adpater,
  if battery into fully charged.
  Check battery LED show white directly when plugging in adpater,
  if battery near full.

Change-Id: I08e6c9ea32dc90599a974d7b5fadaf59f549cc77
Signed-off-by: Devin Lu <Devin.Lu@quantatw.com>
Reviewed-by: Dave Parker <dparker@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203932