stm32f0: Change overslept message to avoid stack overflow
The CPRINTS in the __idle task will cause it to stack overflow.
The reason is that CPRINTS(cprints) calls uart_vprintf and cprintf which also
calls uart_vprintf which calls vfnprintf.

cprints, cprintf both use va_lists. So no wonder that a ~10 character string
with a number in it would cause an overflow given that the format string gets
in the stack twice.

The fix is to call cprintf directly, which does one less layer of va_args.

New StkUsed cap for << idle >> is up to 244/256. It was at 128 before.

BUG=chrome-os-partner:33138
TEST=Comment "if (margin_us < 0)" and note it doesn't crash anymore
BRANCH=None

Change-Id: I1bde52162dac9b8a15332b4b8831862772f4febe
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237020