i2c-stm32f0: Remove sleep while RXing characters
We used to have an usleep(100) that was supposed to context switch while we wait
for bytes or other i2c interrupts.

At 100us it's entirely possible we miss a byte at even 100kHz(90us per byte).
This would be much worse on faster busses. A safe value would have to be under
9us. At that point we wouldn't have time to context switch before we had to
come back.

BUG=chrome-os-partner:33314
TEST=No more stuck I2C transfer because of the EC's fault
TEST=No more SPI Errors
BRANCH=veyron

Change-Id: I9feb598bcb062d4dbbbe4bbeaa6e82baa9288ae7
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/244848
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Alec Berg <alecaberg@chromium.org>