Change ec.bin generation
Previously for the mec1322 chip an ec.bin file was created in the normal
way and then it was "packed" in a post-processing stage to produce
ec.spi.bin. This packing added digital signatures and another loader
stage.

This change allows a chip or board build.mk file to specify the rules
used to produce ec.bin, and uses this for the mec1322 to do the packing.
This means that we use the standard "ec.bin" name, and do not need to
alter other scripts, such as the script which creates
chromeos-firmwareupdate.

BUG=None
BRANCH=NONE
TEST=buildall -j, flash on strago and see it still works.

Change-Id: I3f880d64e60d14f82cb1d21c8b3f2d4ae5e0dfef
Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/259180
Tested-by: Kevin K Wong <kevin.k.wong@intel.com>
Reviewed-by: Shawn N <shawnn@chromium.org>
Commit-Queue: Kevin K Wong <kevin.k.wong@intel.com>