battery: smart: Check Fully charged by BatteryStatus(0x16)
On firmware-veyron branch, for smart battery ec stop charging if
 1. desired_voltage == 0 || desired_current ==0, or
 2. RelativeStateOfCharge == 100%.

In this patch, we use BatteryStatus() instead of RelativeStateOfCharge()
as the fully charged judgment.  Becasue 100% is not exactly 100% charged,
at least on Mighty-rev3 with the newest battery.

Mighty team and Chris@rk also confirmed this is reasonable too.

BUG=chrome-os-partner:34882
BRANCH=veyron
TEST=test on Mighty-rev3 by Mighty team

Change-Id: I645a75aa636fd2dbeee913fc412fa8bd0b8ca2ae
Signed-off-by: ZhengShunQian <zhengsq@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/263660
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Alexandru Stan <amstan@chromium.org>
Commit-Queue: Chris Zhong <zyw@rock-chips.com>
Tested-by: Chris Zhong <zyw@rock-chips.com>