Veyron: All: Adjust battery_info's
Adjust all veyron batteries to match the datasheets
(according to Jerry's spreadsheet).

Mighty has 2 battery voltages, but voltage_max is only really used for
precharging, and it's close enough that it should still work,
otherwise we listen for what the battery actually needs, so it should still
charge to 100%. The alternative was to dynamically switch battery_info the
same way nyan did, which was not worth it.

BUG=None
TEST=Needs a temperature chamber
TEST=Matches the datasheet
BRANCH=veyron

Change-Id: I901e050e3605e307b3f76948c73dd6d4cdfba03e
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/266582
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Queue: Gediminas Ramanauskas <gedis@chromium.org>