Veyron: Don't use timer_arm() directly in the chipset task
It turns out that we really should consider timer_arm() to be a
function only for use by __wait_evt().  If we call it ourselves we can
make things unhappy because we'll trip the "ret == EC_SUCCESS" in
__wait_evt().

All we really needed in the chipset task is something that would wake
us up.  It's really easy to use deferred functions to accomplish that,
so let's do that.

Note: we're up against out limit of deferred functions, but increasing
it doesn't hurt.  ...except on nyan where that blows us over memory,
so we keep it as 8 there.

BRANCH=None
BUG=chrome-os-partner:40942
TEST=Stress the heck out of power button in S0, S3, S5, S7, S9, and S11.
...and S12 too...

Change-Id: I3e1e6d78d03423622eb66eb246555ac0e7fa5674
Suggested-by: Randall Spangler <rspangler@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275071
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Alexandru Stan <amstan@chromium.org>
Tested-by: Alexandru Stan <amstan@chromium.org>