ryu: skip dead battery case when entering RW
When jumping to the RW partition, we don't need to keep the charging
path connected as we do when cold-booting to avoid issue with the dead
battery case.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=smaug
BUG=none
TEST=On Ryu EVT2, plug/unplug a dedicated charger, a USB-PD dual role
device and a C-to-A receptacle with a USB key, check everytime in which
direction the current is flowing with Twinkie, the "AC [on|off]"
messages on the console and the battery charging state.
Do it twice, once with this version flashed as both RO and RW, and once
with a main firmware with soft-sync enabled and an old RO.

Change-Id: Ia2c736435d2e7232efda060b29fdc5a84fdf7d55
Reviewed-on: https://chromium-review.googlesource.com/292969
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>