Ultima: High EC power leakage fix
This patch fixes the EC high power consumption during S3 in
Ultima(BCRD2).
The GPIO201 was configured as OUTPUT and was showing leakage
current in S3 state. This GPIO is not used by EC at all. To fix
it configure it as INPUT to avoid leakage.
This is refer from strago:
https://chromium-review.googlesource.com/292350

BRANCH=None
BUG=None
TEST=Build and boot the ultima board.
Signed-off-by: Keith Tzeng <Keith.Tzeng@quantatw.com>

Change-Id: I2613e1b146266ac6fb3a9ce765ace25cfacaf9d1
Reviewed-on: https://chromium-review.googlesource.com/295812
Reviewed-by: Shawn N <shawnn@chromium.org>
Commit-Queue: Keith Tzeng <keith.tzeng@quantatw.com>
Tested-by: Keith Tzeng <keith.tzeng@quantatw.com>