Fix a power button failure and remove unnecessary flag accesses.
This patch includes the following 2 changes:
1. If the power button is pressed while S5 inactivity timer is about to
   expire, EC need to give CPU a little time to start up before
   changing the state from S5 to G3 (the hard off state); otherwise the
   system will not start up.

2. The invoking of chipset_exit_hard_off() in the power button task is
   always triggered by the tasks whose priorities are lower than the
   chipset task. The other invoking of chipset_exit_hard_off() is done
   by the chipset task itself. Therefore we don't need to worry about
   the situation that chipset_exit_hard_off() is invoked while the
   power state is in POWER_S5G3. Removing unnecessary want_g3_flag
   setting and resetting to let it be better structured.

In addition, we also need to notice that chipset_exit_hard_off() should
be only invoked by the tasks whose priorities are higher than the
chipset task or the chipset task itself, and the trigger of this
invoking should be only done by the tasks whose priorities are lower
than the chipset task or the chipset task itself; otherwise there may
be some concerns of race condition. For example, if we want system to
exit soft-off state for startup, but this function is invoked after S5
inactivity timer expires and before the state changes to POWER_S5G3,
the system will not start up and just go to hard-off state instead.

BUG=chrome-os-partner:42728, chrome-os-partner:42811
BRANCH=None
TEST=Implement an ec command to simulate power button press while S5
     inactivity timer is about to expire, and then make sure that the
     patch did solve the issue.

Change-Id: I548735d45506d82ce0b185c01828d39724dd0b77
Signed-off-by: Kaiyen Chang <kaiyen.chang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/296192
Reviewed-by: Shawn N <shawnn@chromium.org>