ryu: set absolute VINDPM for BQ2589x before changing voltage
When we are transitioning to a lower voltage, we need to ensure that the
under-voltage threshold 'VINDPM' is not set too high else it will
prevent the battery from charging.
So for USB Power Delivery sources which potentially supports multiple
voltages, we set an absolute 'VINDPM' to 4.3V.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=smaug
BUG=chrome-os-partner:45934
TEST=Connect Zinger to Smaug in S5, observe we are charging at 12V,
run the AP, see it downgrading the voltage to 5V, observe the battery
still charges and the BQ25892 REG0D is 0x92.

Change-Id: I93d9a30021355f80be0224ba996ee1496a334e17
Reviewed-on: https://chromium-review.googlesource.com/303369
Reviewed-by: Todd Broch <tbroch@chromium.org>
Trybot-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Alec Berg <alecaberg@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>