ryu: add protection against BQ25892 resetting the current limit
When the BQ25892 is done doing its 'USB' detection, it will reset the
input current limit to the value set by the GPIOs.
At this point, we are getting a charger interrupt, so we set PSEL GPIO to '1'
to limit the current to 100mA until we have time to send the i2c message
to set again the proper current limit.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=smaug
BUG=chrome-os-partner:47972
TEST=plug Apple Macbook Type-C charger and various other chargers and
record the current consumed through Twinkie.

Change-Id: I66f9ebdff5386ce8547bcc009e4ea4b12ae8669a
Reviewed-on: https://chromium-review.googlesource.com/314127
Reviewed-by: Alec Berg <alecaberg@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>