ryu: increase CHARGER task stack size
We got one report from the field of an EC panic due to a stack overflow
in the CHARGER task.
I was not able to reproduce it or find the code path, but as we have free RAM,
we can try to increase it a bit.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=smaug
BUG=b:25977066
TEST=run 'taskinfo' on Smaug and see
'4   CHARGER          00000000    1.103620  408/640'

Change-Id: I443dca4cb2247de5c01b22f30f1e25965376588d
Reviewed-on: https://chromium-review.googlesource.com/315702
Reviewed-by: Alec Berg <alecaberg@chromium.org>
Trybot-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>