Terra: set CONFIG_BATTERY_LEVEL_NEAR_FULL is 96%
The gauge ic of terra is enable charge when rsoc <= 95% and the
BATTERY_LEVEL_NEAR_FULL is 97%, so when the battery is 96%, the charge
LED is amber, but do not charge the battery.

This commit change the threshold of amber led to 96% when
battery rsoc >= 96, the led is green, when battery rsoc < 96,
the led is amber, match the actually charge behavior.

BUG=none
BRANCH=remotes/cros/firmware-strago-7287.B
TEST=Charge full the DUT, keep adapter, wait for rsoc=96%, the charge
led should be green.

Change-Id: I511ec585d6db98f0a9bd2c35a3f7b4969d8d0030
Signed-off-by: james_chao <james_chao@asus.com>
Reviewed-on: https://chromium-review.googlesource.com/330550
Reviewed-by: Shawn N <shawnn@chromium.org>
(cherry picked from commit ca772e8c479cf4a1be011b7e900040f68b6e8024)
Reviewed-on: https://chromium-review.googlesource.com/330580