chell: Indicate when charging in suspend
Currently when in suspend the LED blinks white no matter what the
state of the battery or charging is.  This is very confusing for
users who expect to be able to plug in a charger with the system
in suspend and see that it starts to charge.

Past platforms from this OEM have had two LEDs so this has not
been an issue.

BUG=chrome-os-partner:49151
BRANCH=glados
TEST=put chell in suspend, plug in charger to see amber LED and
then remove the charger and see that it blinks white again.

Change-Id: I60e849d7b8b717fb568d7d5d64046621c1c34157
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/332855