CR50: improve certificate installation error reporting
A lot of things can go wrong when installing certificates, and the
process is happening in a locked down environment, which greatly
complicates troubleshooting.

This patch adds simple error reporting: should there be an error, a
two byte line number of the point where error was reported in
manufacture.c is returned. In case of success, the request phase
returns the full PDU. as before, and the cert installation phase
returns a 16 bit zero value.

An added benefit is receiving of the zero size response, which
indicates that the manufacture command is not supported on the target.

BRANCH=none
BUG=chrome-os-partner:50115
TEST=tried to install a certificate on an unprepared device, observed
     the following expected response:

    $ sudo ./cros_ack.py -i /tmp
    Starting MPSSE at 800 kHz
    Connected to device vid:did:rid of 1ae0:0028:00
    CHIP ID: 00:0200428401000003:010039:0200
    New max timeout: 31 s
    RSA PERSO RESULT: 287
    ECC PERSO RESULT: 341
    FAIL

Change-Id: Ib0661afae322eb70700fd6d12ab2e96017baf0bf
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/360682
Tested-by: Nagendra Modadugu <ngm@google.com>
Reviewed-by: Nagendra Modadugu <ngm@google.com>