i2c_passthru: Return NAK when battery not present.
virtual_battery_read() returns a cached value for some of the supported
smart battery parameters.  If a value isn't supported, it calls out to
the actual battery.  In the case of a battery that's not present, but a
supported battery parameter is queried, we would still return
_something_.  This seems to confuse powerd and causes slow boot.

This commit changes the i2c passthru command to return a NAK when the
virtual battery is enabled but the battery is not present.

BUG=chrome-os-partner:55954
BRANCH=gru
TEST=Build and flash kevin.  Unplug battery.  Verify boot is nice and
quick.

Change-Id: Ib3ab768504c29904f01b91e6194a9689bfdb1e1e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/392926
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Trybot-Ready: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Shawn N <shawnn@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>