common: Print host command code in error.
For some boards, the HCDEBUG setting is defaulting to off due to "too
spammy" console output.  An unfortunate side effect of this is that when
a host command results in an error, it's impossible to know what was the
command that failed.

This commit adds the host command code to the error print so that it's
clear what command failed.

BUG=None
BRANCH=gru,glados,oak
TEST=Build and flash kevin, verify that any host commands that resulted
in an error have the command code printed alongside them.

Change-Id: I6a5f251e7941a47a3cf102a1fb6c5e96ffc8fa5d
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/396127
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>