bob: keep the pull-up on the USB load switch enable
Keep the pull-up enabled all the time on the GPIO connected to the load
switch enable pin.
This is not ideal but the pull-up bit is no-op in output mode on npcx
and this avoids glitching the enable when transition from the 3A mode
(output strong one) to the 1.5A mode (input with a weak pull-up).

That's similar to what we did earlier for Kevin:
https://chromium-review.googlesource.com/411220

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=gru
BUG=b:37694102 chrome-os-partner:59679
TEST=on Bob, plug a USB drive on port 0, then plug another USB device on
port 1 and see that the drive on port 0 is no longer
disconnecting/re-enumerating.

Change-Id: If40cecca250daf897270f6f9d9353661c66ea018
Reviewed-on: https://chromium-review.googlesource.com/498487
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Shasha Zhao <Sarah_Zhao@asus.com>
Tested-by: Shasha Zhao <Sarah_Zhao@asus.com>